--- nutzboot-dubbo-fescar/nutzboot-dubbo-fescar-account/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-server.port=0
-jdbc.url=jdbc:mysql://127.0.0.1/fescar_demo
-jdbc.username=root
-jdbc.password=root
-fescar.enabled=true
-fescar.applicationId=dubbo-fescar-account
-fescar.txServiceGroup=my_test_tx_group
-dubbo.application.name=dubbo-fescar-account
-dubbo.registry.address=zookeeper://127.0.0.1:2181
-dubbo.protocol.name=dubbo
-dubbo.application.qos.enable=false
