+++ nutzboot-dubbo-seata/README.md
@@ -1,0 +1,9 @@
+> nutzboot 升级到最新版 jdk 需要11 以上
+
+
+> dubbo 用的alibaba dubbo 不支持nacos注册中心
+
+> zookeeper
+```bash
+$ docker run -d -p 2181:2181 --name zookeeper zookeeper
+```
