+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-account/src/main/java/io/nutz/demo/dubbo/rpc/DemoSeataAccountServiceLauncher.java
@@ -1,0 +1,30 @@
+package io.nutz.demo.dubbo.rpc;
+
+import io.nutz.demo.bean.Account;
+import org.nutz.boot.NbApp;
+import org.nutz.dao.Cnd;
+import org.nutz.dao.Dao;
+import org.nutz.ioc.loader.annotation.Inject;
+import org.nutz.ioc.loader.annotation.IocBean;
+
+@IocBean(create = "init")
+public class DemoSeataAccountServiceLauncher {
+
+    @Inject
+    Dao dao;
+
+    public void init() {
+        dao.create(Account.class, false);
+        if (dao.count(Account.class, Cnd.where("userId", "=", "U100001")) == 0) {
+            Account account = new Account();
+            account.setUserId("U100001");
+            account.setMoney(1000);
+            dao.insert(account);
+        }
+    }
+
+    public static void main(String[] args) throws Exception {
+        new NbApp().run();
+    }
+
+}
