+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-account/src/main/java/io/nutz/demo/dubbo/rpc/service/impl/AccountServiceImpl.java
@@ -1,0 +1,33 @@
+package io.nutz.demo.dubbo.rpc.service.impl;
+
+import com.alibaba.dubbo.config.annotation.Service;
+import io.seata.core.context.RootContext;
+
+import io.nutz.demo.bean.Account;
+import io.nutz.demo.dubbo.rpc.service.AccountService;
+import org.nutz.dao.Chain;
+import org.nutz.dao.Cnd;
+import org.nutz.dao.Dao;
+import org.nutz.ioc.loader.annotation.Inject;
+import org.nutz.ioc.loader.annotation.IocBean;
+import org.nutz.log.Log;
+import org.nutz.log.Logs;
+
+@IocBean
+@Service(interfaceClass = AccountService.class)
+public class AccountServiceImpl implements AccountService {
+
+    private static final Log log = Logs.get();
+
+    @Inject
+    private Dao dao;
+
+    @Override
+    public void debit(String userId, int money) {
+        log.info("Account Service ... xid: " + RootContext.getXID());
+        log.infof("Deducting balance SQL: update account_tbl set money = money - %s where user_id = %s", money, userId);
+
+        dao.update(Account.class, Chain.makeSpecial("money", "-" + money), Cnd.where("userId", "=", userId));
+        log.info("Account Service End ... ");
+    }
+}
