+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-account/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+server.port=0
+jdbc.url=jdbc:mysql://127.0.0.1/seata_pay
+jdbc.username=root
+jdbc.password=123456
+seata.enabled=true
+seata.applicationId=dubbo-seata-account
+seata.txServiceGroup=my_test_tx_group
+dubbo.application.name=dubbo-seata-account
+dubbo.registry.address=zookeeper://127.0.0.1:2181
+dubbo.protocol.name=dubbo
+dubbo.application.qos.enable=false
