+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-account/src/main/resources/log4j.properties
