+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-common/src/main/java/io/nutz/demo/bean/Stock.java
