+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-order/pom.xml
@@ -1,0 +1,17 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>io.seata</groupId>
+        <artifactId>nutzboot-dubbo-seata</artifactId>
+        <version>1.1.0</version>
+    </parent>
+    <artifactId>nutzboot-dubbo-seata-order</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>io.seata</groupId>
+            <artifactId>nutzboot-dubbo-seata-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
+</project>
