+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-stock/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+server.port=0
+dubbo.application.name=dubbo-seata-stock
+dubbo.registry.address=zookeeper://127.0.0.1:2181
+dubbo.protocol.name=dubbo
+dubbo.application.qos.enable=false
+jdbc.url=jdbc:mysql://127.0.0.1/seata_stock
+jdbc.username=root
+jdbc.password=123456
+seata.enabled=true
+seata.applicationId=dubbo-seata-stock
+seata.txServiceGroup=my_test_tx_group
