+++ nutzboot-dubbo-seata/nutzboot-dubbo-seata-web/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+server.port=8080
+server.host=0.0.0.0
+dubbo.application.name=dubbo-seata-web
+dubbo.registry.address=zookeeper://127.0.0.1:2181
+dubbo.protocol.name=dubbo
+dubbo.application.qos.enable=false
+jdbc.url=jdbc:mysql://127.0.0.1/seata
+jdbc.username=root
+jdbc.password=123456
+seata.enabled=true
+seata.applicationId=dubbo-seata-web
+seata.txServiceGroup=my_test_tx_group
